Only add target offset divs to pages with navbar
For section links on the website (e.g. a link to an element with an ID
such as "neillrobson.com#main"), the previous behavior was to add a
pseudo-element that took up the same amount of width as the fixed navbar
at the top of the screen. For pages without the navbar, this behavior is
unnecessary, so it has been removed and the relevant CSS has been
namespaced under a class.